Simplify logger structure
This commit simplifies the logger structure of each pattern by having a single logger instance with the package name of the pattern which lives in the `App.kt` file of each pattern and have the `internal` exposure level. Hence, only classes within the pattern package can use it, and everything logged in the pattern appears with the same prefix which makes it easier to read.